+++ flash/core/data/transforms.py
@@ -78,16 +78,21 @@ class KorniaParallelTransforms(nn.Sequential):
         result = list(inputs) if isinstance(inputs, Sequence) else [inputs]
         for transform in self.children():
             inputs = result
+
+            # we enforce the first time to sample random params
+            result[0] = transform(inputs[0])
+
             if hasattr(transform, "_params") and bool(transform._params):
                 params = transform._params
             else:
                 params = None
 
-            for i, input in enumerate(inputs):
+            # apply transforms from (1, n)
+            for i, input in enumerate(inputs[1:]):
                 if params is not None:
-                    result[i] = transform(input, params)
+                    result[i + 1] = transform(input, params)
                 else:  # case for non random transforms
-                    result[i] = transform(input)
+                    result[i + 1] = transform(input)
             if hasattr(transform, "_params") and bool(transform._params):
                 transform._params = None
         return result